feat: skeleton 적용"
